added CUDA
Former-commit-id: 21ec1eef2fa21aaea341e228f8b19fb5cf43ae62

+++ src/object-log/BoundingBox.cpp
@@ -1,13 +1,13 @@
 #include "BoundingBox.h"
 
-BoundingBox::BoundingBox(double xCenter, double yCenter, double depthVal, double width, double height)
+BoundingBox::BoundingBox(float xCenter, float yCenter, float depthVal, float width, float height)
 {
     _xCenter = xCenter;
     _yCenter = yCenter;
     _depthVal = depthVal;
     _width = width;
     _height = height;
-    _position = std::tuple<double, double, double>((xCenter - (width/2)), yCenter - (height/2), depthVal);
+    _position = std::tuple<float, float, float>((xCenter - (width/2)), yCenter - (height/2), depthVal);
     _timeStamp = time(0);
     
 }
@@ -18,31 +18,31 @@ BoundingBox::BoundingBox(){
     _depthVal = 0;
     _width = 0;
     _height = 0;
-    _position = std::tuple<double, double, double>(0, 0, 0);
+    _position = std::tuple<float, float, float>(0, 0, 0);
     _timeStamp = time(0);
 }
 
-void BoundingBox::setXCenter(double xCenter){
+void BoundingBox::setXCenter(float xCenter){
     _xCenter = xCenter;
 }
 
-void BoundingBox::setYCenter(double yCenter){
+void BoundingBox::setYCenter(float yCenter){
     _yCenter = yCenter;
 }
 
-void BoundingBox::setDepthVal(double depthVal){
+void BoundingBox::setDepthVal(float depthVal){
     _depthVal = depthVal;
 }
 
-void BoundingBox::setWidth(double width){
+void BoundingBox::setWidth(float width){
     _width = width;
 }
 
-void BoundingBox::setHeight(double height){
+void BoundingBox::setHeight(float height){
     _height = height;
 }
 
-void BoundingBox::setPosition(std::tuple<double, double, double> position){
+void BoundingBox::setPosition(std::tuple<float, float, float> position){
     _position = position;
 }
 
@@ -50,27 +50,27 @@ void BoundingBox::setTimeStamp(time_t timeStamp){
     _timeStamp = timeStamp;
 }
 
-double BoundingBox::getXCenter(){
+float BoundingBox::getXCenter(){
     return _xCenter;
 }
 
-double BoundingBox::getYCenter(){
+float BoundingBox::getYCenter(){
     return _yCenter;
 }
 
-double BoundingBox::getDepthVal(){
+float BoundingBox::getDepthVal(){
     return _depthVal;
 }
 
-double BoundingBox::getWidth(){
+float BoundingBox::getWidth(){
     return _width;
 }
 
-double BoundingBox::getHeight(){
+float BoundingBox::getHeight(){
     return _height;
 }
 
-std::tuple<double, double, double> BoundingBox::getPosition(){
+std::tuple<float, float, float> BoundingBox::getPosition(){
     return _position;
 }
